chore: move tests and create sow script
