+++ .gitignore
@@ -13,3 +13,6 @@
 
 # generated by CI
 init-args.json
+
+# Scope-of-work
+sow.txt
