+++ common/src/collection/chunked_append_only_list.rs
@@ -1,6 +1,9 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 use near_sdk::{env, near, store::Vector, BorshStorageKey, IntoStorageKey};
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Debug, Clone, Copy, BorshSerialize, BorshStorageKey, PartialEq, Eq, PartialOrd, Ord)]
 enum StorageKey {
     Inner,
@@ -154,67 +157,5 @@ impl<T: BorshSerialize + BorshDeserialize, const CHUNK_SIZE: u32> DoubleEndedIte
             .and_then(|index| self.list.get(index).map(|x| (index, x)))?;
         self.until_index = index;
         Some(value)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn basic() {
-        let mut list = ChunkedAppendOnlyList::<_, 47>::new(b"l");
-        assert_eq!(list.len(), 0);
-        assert!(list.is_empty());
-
-        for i in 0..10_000usize {
-            list.push(i);
-            assert_eq!(list.len() as usize, i + 1);
-            assert!(!list.is_empty());
-        }
-
-        let mut count = 0;
-        for (i, v) in list.iter().enumerate() {
-            assert_eq!(i, *v);
-            count += 1;
-        }
-
-        assert_eq!(count, 10_000);
-    }
-
-    #[test]
-    fn replace_last() {
-        let mut list = ChunkedAppendOnlyList::<_, 47>::new(b"l");
-        for i in 0..10_000u32 {
-            list.push(i);
-            list.replace_last(i * 2);
-            assert_eq!(list.len(), i + 1);
-            assert!(!list.is_empty());
-        }
-
-        for i in 0..10_000u32 {
-            let x = list.get(i).unwrap();
-            assert_eq!(*x, i * 2);
-        }
-
-        assert_eq!(list.len(), 10_000);
-    }
-
-    #[test]
-    fn next_back() {
-        let mut list = ChunkedAppendOnlyList::<_, 47>::new(b"l");
-        for i in 0..10_000u32 {
-            list.push(i);
-        }
-
-        let mut it = list.iter();
-
-        let mut i = 10_000;
-        while let Some(x) = it.next_back() {
-            i -= 1;
-            assert_eq!(*x, i);
-        }
-
-        assert_eq!(i, 0);
     }
 }
