+++ common/src/data/accumulator.rs
@@ -2,6 +2,9 @@ use near_sdk::{json_types::U128, near, require};
 
 use crate::asset::{AssetClass, FungibleAssetAmount};
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[near(serializers = [borsh, json])]
 pub struct Accumulator<T: AssetClass> {
@@ -106,56 +109,5 @@ pub struct AccumulationRecord<T: AssetClass> {
 impl<T: AssetClass> AccumulationRecord<T> {
     pub fn get_amount(&self) -> FungibleAssetAmount<T> {
         self.amount
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn amortization() {
-        let mut a = Accumulator::<crate::asset::BorrowAsset>::new(1);
-
-        a.accumulate(AccumulationRecord {
-            amount: 100.into(),
-            fraction_as_u128_dividend: 0,
-            next_snapshot_index: 2,
-        });
-
-        assert_eq!(a.get_total(), 100.into());
-
-        a.amortize(25.into());
-
-        assert_eq!(a.get_total(), 125.into());
-
-        a.accumulate(AccumulationRecord {
-            amount: 100.into(),
-            fraction_as_u128_dividend: 0,
-            next_snapshot_index: 3,
-        });
-
-        assert_eq!(a.get_total(), 200.into());
-    }
-
-    #[test]
-    fn fraction() {
-        let mut a = Accumulator::<crate::asset::BorrowAsset>::new(1);
-
-        a.accumulate(AccumulationRecord {
-            amount: 100.into(),
-            fraction_as_u128_dividend: 1 << 127,
-            next_snapshot_index: 2,
-        });
-
-        assert_eq!(a.get_total(), 100.into());
-
-        a.accumulate(AccumulationRecord {
-            amount: 100.into(),
-            fraction_as_u128_dividend: 1 << 127,
-            next_snapshot_index: 3,
-        });
-
-        assert_eq!(a.get_total(), 201.into());
     }
 }
