+++ common/src/data/snapshot.rs
