+++ common/src/data/static_yield.rs
