+++ common/src/market/event.rs
