+++ common/src/market/impl.rs
@@ -10,8 +10,7 @@ use crate::{
     asset_op,
     borrow::{BorrowPosition, BorrowPositionGuard, BorrowPositionRef},
     chunked_append_only_list::ChunkedAppendOnlyList,
-    event::MarketEvent,
-    market::{MarketConfiguration, WithdrawalResolution},
+    market::{event::MarketEvent, MarketConfiguration, WithdrawalResolution},
     number::Decimal,
     snapshot::Snapshot,
     static_yield::StaticYieldRecord,
